+++ client/src/Pages/Checks/Components/OptionsHeader/index.jsx
@@ -5,6 +5,7 @@ import PropTypes from "prop-types";
 
 //Utils
 import { useTheme } from "@emotion/react";
+import SkeletonLayout from "./skeleton.jsx";
 import { useTranslation } from "react-i18next";
 
 const OptionsHeader = ({
@@ -14,27 +15,27 @@ const OptionsHeader = ({
 	monitors,
 	filter = "all",
 	setFilter,
-	dateRange = "all",
+	dateRange = "hour",
 	setDateRange,
 }) => {
 	const theme = useTheme();
 	const { t } = useTranslation();
 	const monitorNames = typeof monitors !== "undefined" ? Object.values(monitors) : [];
-
 	const filterOptions = [
-		{ _id: "all", name: t("incidentsPage.incidentsOptionsHeaderFilterAll") },
-		{ _id: "active", name: t("incidentsPage.incidentsOptionsHeaderFilterActive") },
-		{ _id: "resolved", name: t("incidentsPage.incidentsOptionsHeaderFilterResolved") },
-		{ _id: "manual", name: t("incidentsPage.incidentsOptionsHeaderFilterManual") },
+		{ _id: "all", name: t("incidentsOptionsHeaderFilterAll") },
+		{ _id: "down", name: t("incidentsOptionsHeaderFilterDown") },
+		{ _id: "resolve", name: t("incidentsOptionsHeaderFilterCannotResolve") },
+		{ _id: "resolved", name: t("incidentsOptionsHeaderFilterResolved") },
 	];
 
+	// The stacks below which are three in number have the same style so
 	const stackStyles = {
 		direction: "row",
 		alignItems: "center",
 		gap: theme.spacing(6),
 	};
 
-	if (!shouldRender) return;
+	if (!shouldRender) return <SkeletonLayout />;
 
 	return (
 		<Stack
