+++ client/src/Pages/Checks/Components/StatusBoxes/skeleton.jsx
