+++ client/src/Pages/Incidents/Components/StatisticsPanel/index.jsx
